refactor(app/game): refactored code - App handle game start and take args for reproducible run
